+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.2-dev
 
+- `Input.value` to provide the current value in the `<input>` field.
+
 ## 0.0.1
 
 - Initial version.